+++ weblate/trans/views/acl.py
@@ -134,7 +134,8 @@ def block_user(request: AuthenticatedHttpRequest, project):
         else:
             expiry = None
         _userblock, created = user.userblock_set.get_or_create(
-            project=obj, defaults={"expiry": expiry}
+            project=obj,
+            defaults={"expiry": expiry, "note": form.cleaned_data.get("note", "")},
         )
         if created:
             AuditLog.objects.create(